+++ internal/suite/setup.go
@@ -20,7 +20,7 @@ import (
 // Setup handles the creation of the BeforeSuite and AfterSuite handlers. This covers the creations and cleanup of the test cluster.
 // `clusterReadyFns` can be provided if the cluster requires custom checks for cluster-ready status. If not provided the cluster will
 // be checked for at least a single control plane node being marked as ready.
-func Setup(isUpgrade bool, kubeContext string, clusterBuilder cb.ClusterBuilder, clusterReadyFns ...func(client *client.Client)) {
+func Setup(isUpgrade bool, clusterBuilder cb.ClusterBuilder, clusterReadyFns ...func(client *client.Client)) {
 	BeforeSuite(func() {
 		if isUpgrade && strings.TrimSpace(os.Getenv("E2E_OVERRIDE_VERSIONS")) == "" {
 			Skip("E2E_OVERRIDE_VERSIONS env var not set, skipping upgrade test")
@@ -31,7 +31,7 @@ func Setup(isUpgrade bool, kubeContext string, clusterBuilder cb.ClusterBuilder,
 
 		state.SetContext(context.Background())
 
-		framework, err := clustertest.New(kubeContext)
+		framework, err := clustertest.New(clusterBuilder.KubeContext())
 		Expect(err).NotTo(HaveOccurred())
 		state.SetFramework(framework)